Moved VertxContextScope to correct package
