+++ src/main/java/org/prebid/server/spring/config/ServiceConfiguration.java
@@ -22,7 +22,6 @@ import org.prebid.server.gdpr.GdprService;
 import org.prebid.server.metric.Metrics;
 import org.prebid.server.optout.GoogleRecaptchaVerifier;
 import org.prebid.server.settings.ApplicationSettings;
-import org.prebid.server.spring.VertxContextScope;
 import org.prebid.server.validation.BidderParamValidator;
 import org.prebid.server.validation.RequestValidator;
 import org.prebid.server.validation.ResponseBidValidator;
